@@ -1,16 +1,16 @@
 ﻿#pragma once
 #include <string>
+#include <memory>
 #include <iostream>
 
-#include "Solver.h"
+#include "Logics/Solver.h"
 
 class Calculator {
 public:
   Calculator(int argc = 0, char* argv[] = nullptr);
-
   int exec(std::istream& is = std::cin, std::ostream& os = std::cout);
 
 private:  
   std::string pluginsPath;
-  Solver solver;
+  std::unique_ptr<Solver> solver;
 };
