@@ -1,20 +1,21 @@
-#include "../Operators/Operator.h"
+#define _USE_MATH_DEFINES
 #include <cmath>
-#define M_PI 3.141592653589793
 
-class Cos : public Operator {
+#include "../Operators/Operator.h"
+
+class Cos : public Function {
 public:
-  double operator()() const override { return std::cos(args[0] / 180.0 * M_PI); }
-  EPriority getPriority() const override { return EPriority::P_FUNCTION; }
+  double operator()() const override { return std::cos((args[0] * M_PI) / 180.0); }
   std::string getName() const override { return "cos"; }
+  std::string getDescription() const override { return "cos(deg)"; }
 
-  bool isFunction() const override { return true; }
   size_t getnArgs() const override { return 1; }
   void setArg(size_t narg, double arg) override { args[narg] = arg; }
 private:
   double args[1] = { 0 };
 };
 
-extern "C" __declspec(dllexport) Operator * create() {
+extern "C" __declspec(dllexport) 
+Operator * create() {
   return new Cos();
 }
